CodeVeinFix: Added new way of correcting FOV on different resolutions
Added new submodules.

	modified:   .gitmodules

Added new dependencies into build.

	modified:   CMakeLists.txt

Changed the way that FOV is corrected. Previously
FOV was changed by playing around with the hex
value 0x3c0efa35 or 0.00872665. This value is
pi/360 which is part of the equation that
calculates the FOV. The games default FOV is 68,
with this knowledge it was possible to now use
the FOV equation and calculate the FOV based on
desktop resolution or the resolution provided by
the user in yml file.

	modified:   src/main.cpp