CodeVeinFix: Massive overhaul
FOV is no longer a multiplier, but a flat value.
The default flat value is whatever the game
default is. This can then be customized by the
user.

LOG file will now have compiler information to
assist with debugging.

	modified:   src/main.cpp

Added new function which will get compiler info,
this is to be used to help with debugging.

	modified:   src/utils.cpp
	modified:   inc/utils.hpp

Release builds for Github shall now be made with
GCC and will staticly link libgcc and libstdc++.
Using GCC and those libs will prevent user errors
with MSVC where certain DLLs are missing which
will cause the fix to not load and subsequently
not load the game. This has been made into a
cmake build option, those wishing to build from
source can still use any compiler and any build
system, but for release builds on Github GCC
builds will be forced to prevent missing DLL
issues as stated previously.

Also made the binaries go into their own folder
as opposed to the build folder with other cmake
artifacts for convenience.

	modified:   CMakeLists.txt

Added bin folder to ignore list so binaries are
not accidently uploaded Github repo.

	modified:   .gitignore

Fixed a bug where the script was not using the
passed in cli arguments.

Added usage page and cli checker utility to script
to prevent undefined behavior on bad args.

The yml script has been changed to accomodate FOV
changes in the code above.

And lastly changed where the script will attempt
to look for dll file. Note it will recursively
traverse the bin folder and use the first dll it
finds regardless if its the debug or release dll.
This is by design and NOT a bug.

	modified:   install.ps1

submodule stuff, no idea why this got picked up...

	new file:   safetyhook
	new file:   zydis